requirements.txt
Flask==3.0.0
python-dotenv==1.0.1
openai>=1.35.12
pdfkit==1.0.0

